# roc-models/sca-0.1.x/openapi.yaml
...
components:
  parameters:
    city-id:
      content:
        text/plain; charset=utf-8:
          schema:
            type: string
      description: city-id (target in onos-config)
      in: path
      name: city-id
      required: true
...
  schemas:
...
    AdditionalPropertyCity-Id:
      description: Optionally specify a city-id other than the default (only on PATCH
        method)
      properties:
        city-id:
          description: an override of the city-id (target)
          title: city-id
          type: string
      title: AdditionalPropertyCity-Id
      type: object
    AdditionalPropertyUnchanged:
      description: To optionally omit 'required' properties, add them to 'unchanged'
        list
      properties:
        unchanged:
          description: A comma seperated list of unchanged mandatory attribute names
          title: unchanged
          type: string
      title: AdditionalPropertyUnchanged
      type: object
...
    District:
      additionalProperties:
        $ref: '#/components/schemas/AdditionalPropertyCity-Id'
      description: district is a subsection of a city (single)
      properties:
        description:
          description: long description field
          maxLength: 1024
          minLength: 1
          title: description
          type: string
        display-name:
          description: display name to use in GUI or CLI
          maxLength: 80
          minLength: 1
          title: display-name
          type: string
        district-id:
          description: An identifier for the district
          maxLength: 80
          minLength: 1
          title: district-id
          type: string
          x-go-type: ListKey
        image:
          description: a URL to an image file (e.g. SVG, PNG). Optional
          maxLength: 200
          minLength: 1
          title: image
          type: string
        location:
          $ref: '#/components/schemas/District_Location'
        source:
          $ref: '#/components/schemas/District_Source_List'
      required:
      - district-id
      title: District
      type: object
      x-list-multiple: true
    District_List:
      description: district is a subsection of a city (list)
      items:
        $ref: '#/components/schemas/District'
      type: array
      uniqueItems: true
      x-keys:
      - district-id
      x-list-multiple: true
    District_Location:
      additionalProperties:
        $ref: '#/components/schemas/AdditionalPropertyUnchanged'
      description: physical location in grid or latitude and longitude
      properties:
        coordinate-system:
          description: choice of coordinate system
          enum:
          - grid-bottom-left
          - grid-top-left
          - lat-long
          title: coordinate-system
          type: string
        orientation-x:
          description: rotation in dgrees the X-Y plane clockwise from the 3 o'clock
            position
          format: int64
          title: orientation-x
          type: integer
        x:
          description: The X position in grid or longitude by 10e-7
          format: int64
          title: x
          type: integer
        "y":
          description: The Y position in grid or latitude by 10e-7
          format: int64
          title: "y"
          type: integer
        z:
          description: The Z position in grid or elevation in centimetres (optional)
          format: int64
          title: z
          type: integer
      required:
      - coordinate-system
      - x
      - "y"
      title: District_Location
      type: object
    District_Source:
      description: The list of cameras (single)
      properties:
        description:
          description: long description field
          maxLength: 1024
          minLength: 1
          title: description
          type: string
        display-name:
          description: display name to use in GUI or CLI
          maxLength: 80
          minLength: 1
          title: display-name
          type: string
        image:
          description: a URL to an image file (e.g. SVG, PNG). Optional
          maxLength: 200
          minLength: 1
          title: image
          type: string
        location:
          $ref: '#/components/schemas/District_Source_Location'
        source-id:
          description: the name of the inference source
          title: source-id
          type: string
          x-go-type: ListKey
        state:
          $ref: '#/components/schemas/District_Source_State'
        video:
          $ref: '#/components/schemas/District_Source_Video'
      required:
      - source-id
      title: District_Source
      type: object
      x-list-multiple: true
    District_Source_List:
      description: The list of cameras (list)
      items:
        $ref: '#/components/schemas/District_Source'
      type: array
      uniqueItems: true
      x-keys:
      - source-id
      x-list-multiple: true
    District_Source_Location:
      additionalProperties:
        $ref: '#/components/schemas/AdditionalPropertyUnchanged'
      description: physical location in grid or latitude and longitude
      properties:
        coordinate-system:
          description: choice of coordinate system
          enum:
          - grid-bottom-left
          - grid-top-left
          - lat-long
          title: coordinate-system
          type: string
        orientation-x:
          description: rotation in dgrees the X-Y plane clockwise from the 3 o'clock
            position
          format: int64
          title: orientation-x
          type: integer
        x:
          description: The X position in grid or longitude by 10e-7
          format: int64
          title: x
          type: integer
        "y":
          description: The Y position in grid or latitude by 10e-7
          format: int64
          title: "y"
          type: integer
        z:
          description: The Z position in grid or elevation in centimetres (optional)
          format: int64
          title: z
          type: integer
      required:
      - coordinate-system
      - x
      - "y"
      title: District_Source_Location
      type: object
    District_Source_State:
      description: read only attributes of the source - populated by adapter from
        analytics engine
      properties:
        error:
          description: description of error if present
          readOnly: true
          title: error
          type: string
        error-since:
          description: date and time error last updated
          readOnly: true
          title: error-since
          type: string
        status:
          description: status of the source
          readOnly: true
          title: status
          type: string
      title: District_Source_State
      type: object
    District_Source_Video:
      additionalProperties:
        $ref: '#/components/schemas/AdditionalPropertyUnchanged'
      description: attributes related to video stream
      properties:
        path:
          description: the path to the inference source
          maxLength: 100
          minLength: 1
          pattern: '[a-zA-Z0-9\./_:-]*'
          title: path
          type: string
        source-type:
          description: Source type
          enum:
          - device
          - file
          - sample
          - stream
          title: source-type
          type: string
      required:
      - path
      - source-type
      title: District_Source_Video
      type: object
...
info:
  contact:
    name: Intel Corporation
    url: https://www.intel.com
  description: OpenAPI 3 specification is generated from sca onos-config model plugin
  license:
    name: LicenseRef-Intel
  title: sca-0.1.x
  version: 0.1.x
openapi: 3.0.0
paths:
...
  /sca/v0.1.x/{city-id}/district:
    description: district is a subsection of a city
    get:
      operationId: getDistrict_List
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/District_List'
          description: GET OK 200
      summary: GET /district List
    parameters:
    - $ref: '#/components/parameters/city-id'
...